+++ bat/release-ios.sh
@@ -1,3 +1,11 @@
+if [ -z $2 ]; then
+  echo 'Error. Please enter <Alias>'
+  exit
+fi
+
+# change config.xml language
+node ./helper/config.js --lang $1
+
 rm ./apk/ios-signed.apk
 rm ./apk/ios-release.apk
 
@@ -6,7 +14,7 @@ rm -r ./platforms
 cordova platform add ios
 
 cordova build android --release
-jarsigner -verbose -keystore .keystore platforms/ios/app/build/outputs/apk/release/app-release-unsigned.apk $1
+jarsigner -verbose -keystore .keystore platforms/ios/app/build/outputs/apk/release/app-release-unsigned.apk $2
 cp ./platforms/ios/app/build/outputs/apk/release/app-release-unsigned.apk ./apk/
 
 mv ./apk/app-release-unsigned.apk ./apk/ios-signed.apk
